requirements.txt
mcp
anthropic
python-dotenv
flask